% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc-releases.R
\name{wood_bioc_releases}
\alias{wood_bioc_releases}
\title{List Bioconductor releases}
\usage{
wood_bioc_releases()
}
\value{
A character vector, where each element is a Bioconductor release.
}
\description{
This function scrapes the Bioconductor site for version codes
of all its releases.
}
\examples{
\dontshow{if (!woodendesc:::is_cran_check()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
wood_bioc_releases()
\dontshow{\}) # examplesIf}
}
\seealso{
Functions for Bioconductor:
\code{\link{wood_bioc_dependencies}()},
\code{\link{wood_bioc_packages}()},
\code{\link{wood_bioc_version}()}
}
\concept{bioc}
