% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{wood_clear_cache}
\alias{wood_clear_cache}
\title{Clear cache}
\usage{
wood_clear_cache()
}
\value{
An integer code, invisibly, the same as in \code{\link[base:unlink]{base::unlink()}}.
}
\description{
Remove all \pkg{woodendesc}-related cached files.
}
