% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-packages.R
\name{wood_core_packages}
\alias{wood_core_packages}
\title{List core R packages}
\usage{
wood_core_packages()
}
\value{
A character vector of available packages.
}
\description{
This function returns the list of base R packages.
}
\examples{
wood_core_packages()

}
\seealso{
Functions for core R packages:
\code{\link{wood_core_dependencies}()},
\code{\link{wood_core_version}()}

Functions that query available packages:
\code{\link{wood_bioc_packages}()},
\code{\link{wood_cran_packages}()},
\code{\link{wood_github_packages}()},
\code{\link{wood_local_packages}()},
\code{\link{wood_packages}()},
\code{\link{wood_runiverse_packages}()},
\code{\link{wood_url_packages}()}
}
\concept{core}
\concept{packages}
