% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-latest.R
\name{wood_cran_latest}
\alias{wood_cran_latest}
\title{Get latest package version on CRAN}
\usage{
wood_cran_latest(package)
}
\arguments{
\item{package}{[\code{character(1)}]\cr
Name of a package to query.}
}
\value{
A single string with a version code.
}
\description{
This function queries CRAN for the code of the latest version
of the selected package.
}
\examples{
\donttest{
wood_cran_latest("versionsort")
}

}
\concept{cran}
\concept{versions}
