% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-packages.R
\name{wood_cran_packages}
\alias{wood_cran_packages}
\title{List available packages on CRAN}
\usage{
wood_cran_packages()
}
\value{
A character vector of available packages.
}
\description{
This function queries CRAN for a list of available packages.
They are returned as a vector of strings, each element being a package name.
}
\examples{
\donttest{
head(wood_cran_packages())
}

}
\seealso{
Functions for CRAN:
\code{\link{wood_cran_dependencies}()},
\code{\link{wood_cran_latest}()},
\code{\link{wood_cran_versions}()}

Functions that query available packages:
\code{\link{wood_bioc_packages}()},
\code{\link{wood_core_packages}()},
\code{\link{wood_github_packages}()},
\code{\link{wood_local_packages}()},
\code{\link{wood_packages}()},
\code{\link{wood_runiverse_packages}()},
\code{\link{wood_url_packages}()}
}
\concept{cran}
\concept{packages}
