% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab-dependencies.R
\name{wood_gitlab_dependencies}
\alias{wood_gitlab_dependencies}
\title{Get dependencies of a package on GitLab}
\usage{
wood_gitlab_dependencies(package, user, tag = "latest")
}
\arguments{
\item{package}{\code{character(1)}\cr
Name of a package to query for.}

\item{user}{\code{character(1)}\cr
Name of a user or organization.}

\item{tag}{\code{character(1)}\cr
Tag of a commit on GitLab or \code{"latest"} for the latest (possibly untagged)
commit.}
}
\value{
A data frame with three columns, all in string format:
\itemize{
\item \code{package} (package name),
\item \code{version} (minimum version requirement or \code{NA} if none),
\item \code{type} (dependency type, e.g. \code{"Imports"}).
}
}
\description{
This function queries GitLab for dependencies of the selected
tagged commit of a repo. By default, it queries the latest commit instead.
}
\examples{
## Only run with internet access
if (interactive()) {
wood_gitlab_dependencies("limonaid", "r-packages")
}

if (interactive()) {
wood_gitlab_dependencies("rock", "r-packages", tag = "0.6.3")
}

}
\seealso{
Functions for GitLab:
\code{\link{wood_gitlab_latest}()},
\code{\link{wood_gitlab_packages}()},
\code{\link{wood_gitlab_tags}()},
\code{\link{wood_gitlab_versions}()}

Functions that query package dependencies:
\code{\link{wood_bioc_dependencies}()},
\code{\link{wood_core_dependencies}()},
\code{\link{wood_cran_dependencies}()},
\code{\link{wood_dependencies}()},
\code{\link{wood_github_dependencies}()},
\code{\link{wood_local_dependencies}()},
\code{\link{wood_runiverse_dependencies}()},
\code{\link{wood_url_dependencies}()}
}
\concept{dependencies}
\concept{gitlab}
