% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab-packages.R
\name{wood_gitlab_packages}
\alias{wood_gitlab_packages}
\title{List available package on a Gitlab account}
\usage{
wood_gitlab_packages(user, include_forks = FALSE)
}
\arguments{
\item{user}{\code{character(1)}\cr
Name of a user or organization.}

\item{include_forks}{\code{logical(1)}\cr
Whether to include packages forked from other accounts.}
}
\value{
A character vector of available packages.
}
\description{
This function finds packages among repositories belonging to a
selected account. They are returned as a vector of strings, each element
being a repository (and in most cases, package) name.
}
\examples{
if (interactive()) {
  wood_gitlab_packages("r-packages")
  # The function takes care of differentiating
  # between users and groups internally
  wood_gitlab_packages("matherion")
}

}
\seealso{
Functions for GitLab:
\code{\link{wood_gitlab_dependencies}()},
\code{\link{wood_gitlab_latest}()},
\code{\link{wood_gitlab_tags}()},
\code{\link{wood_gitlab_versions}()}

Functions that query available packages:
\code{\link{wood_bioc_packages}()},
\code{\link{wood_core_packages}()},
\code{\link{wood_cran_packages}()},
\code{\link{wood_github_packages}()},
\code{\link{wood_local_packages}()},
\code{\link{wood_packages}()},
\code{\link{wood_runiverse_packages}()},
\code{\link{wood_url_packages}()}
}
\concept{gitlab}
\concept{packages}
