% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab-tags.R
\name{wood_gitlab_tags}
\alias{wood_gitlab_tags}
\title{List GitLab tags for a repository}
\usage{
wood_gitlab_tags(package, user)
}
\arguments{
\item{package}{\code{character(1)}\cr
Name of a package to query for.}

\item{user}{\code{character(1)}\cr
Name of a user or organization.}
}
\value{
A character vector of repository tags.
}
\description{
This function queries GitLab for a list of tags for a given
repository (indicated by a combination of package name and username). Each
element is a tag, which need not be identical or even related to version code
of the package at the given time.
}
\examples{
\donttest{
wood_gitlab_tags("rock", "r-packages")
}

}
\seealso{
Functions for GitLab:
\code{\link{wood_gitlab_dependencies}()},
\code{\link{wood_gitlab_latest}()},
\code{\link{wood_gitlab_packages}()},
\code{\link{wood_gitlab_versions}()}
}
\concept{gitlab}
