% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab-versions.R
\name{wood_gitlab_versions}
\alias{wood_gitlab_versions}
\title{Get all tagged package versions on GitLab}
\usage{
wood_gitlab_versions(package, user)
}
\arguments{
\item{package}{\code{character(1)}\cr
Name of a package to query for.}

\item{user}{\code{character(1)}\cr
Name of a user or organization.}
}
\value{
A character vector of version codes.
}
\description{
This function queries GitLab for the codes of all tagged package
versions. To get the latest available version, see \code{\link[=wood_gitlab_latest]{wood_gitlab_latest()}}.
}
\examples{
\dontshow{if (!woodendesc:::is_cran_check()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
wood_gitlab_versions("rock", "r-packages")
\dontshow{\}) # examplesIf}
}
\seealso{
Functions for GitLab:
\code{\link{wood_gitlab_dependencies}()},
\code{\link{wood_gitlab_latest}()},
\code{\link{wood_gitlab_packages}()},
\code{\link{wood_gitlab_tags}()}

Functions that query package versions:
\code{\link{wood_bioc_version}()},
\code{\link{wood_core_version}()},
\code{\link{wood_cran_latest}()},
\code{\link{wood_cran_versions}()},
\code{\link{wood_github_latest}()},
\code{\link{wood_github_versions}()},
\code{\link{wood_gitlab_latest}()},
\code{\link{wood_local_versions}()},
\code{\link{wood_runiverse_version}()},
\code{\link{wood_url_version}()},
\code{\link{wood_versions}()}
}
\concept{gitlab}
\concept{versions}
