% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-packages.R
\name{wood_local_packages}
\alias{wood_local_packages}
\title{List locally available packages}
\usage{
wood_local_packages(path = .libPaths()[1])
}
\arguments{
\item{path}{[\code{character(1)}]\cr
A path to local library, by default the first in \code{.libPaths()}.}
}
\value{
A character vector of available packages.
}
\description{
This function searches for installed packages inside a specified
library path.
}
\examples{
\donttest{
wood_local_packages()
}

}
\seealso{
Functions that query available packages:
\code{\link{wood_cran_packages}()},
\code{\link{wood_runiverse_packages}()},
\code{\link{wood_url_packages}()}
}
\concept{local}
\concept{packages}
