% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-packages.R
\name{wood_local_packages}
\alias{wood_local_packages}
\title{List locally available packages}
\usage{
wood_local_packages(paths = .libPaths()[1])
}
\arguments{
\item{paths}{\code{character()}\cr
Paths to local libraries, by default the first element of \code{.libPaths()}. If
equal to \code{"all"}, uses all elements of \code{.libPaths()}.}
}
\value{
A character vector of available packages.
}
\description{
This function searches for installed packages inside specified
library paths.
}
\examples{
wood_local_packages()
wood_local_packages("all")

}
\seealso{
Functions for local libraries:
\code{\link{wood_local_dependencies}()},
\code{\link{wood_local_versions}()}

Functions that query available packages:
\code{\link{wood_bioc_packages}()},
\code{\link{wood_core_packages}()},
\code{\link{wood_cran_packages}()},
\code{\link{wood_github_packages}()},
\code{\link{wood_packages}()},
\code{\link{wood_runiverse_packages}()},
\code{\link{wood_url_packages}()}
}
\concept{local}
\concept{packages}
